Extract embed query helpers from resize broadcast module.
Move shouldEmbed and shouldAutoplay into embedMode so measurement code
and play-function policy live in separate modules.
